Fix the randomness output of the model.eval()
As the origin git repo, there is a random output of the model.eval() which
makes the experiments could not be reproduced. However, the root cause is
the difference between `nn.Dropout()` and `F.Dropout()`. After changing
the `F.Dropout()` to `nn.Dropout()`, the result of evaluation become *static*.
Further examination is in `inference.ipynb`.